Fuse fmpz_mpoly/divides*
